// components/simpleSelect/targets/simpleSelectTargets.enum.ts
export enum SimpleSelectTargets {
  label = 'label',
  control = 'control',
  controlWrapper = 'controlWrapper',
  popper = 'popper',
  readOnly = 'readOnly',
  root = 'root',
  selectedValue = 'selectedValue',
  adornments = 'adornments',
  adornmentsLeft = 'adornmentsLeft',
  adornmentsRight = 'adornmentsRight',
  errorsRoot = 'errorsRoot',
  errorsUnorderedList = 'errorsUnorderedList',
  errorsUnorderedListItem = 'errorsUnorderedListItem',
  errorsFieldError = 'errorsFieldError',
  /** Simple View Only */
  errorsTooltip = 'errorsTooltip',
  errorsIcon = 'errorsIcon',
}